updated status.js file
